+++ packages/multichain-network-controller/src/index.ts
@@ -1,0 +1,24 @@
+export { MultichainNetworkController } from './MultichainNetworkController';
+export { getDefaultMultichainNetworkControllerState } from './constants';
+export type {
+  MultichainNetworkMetadata,
+  SupportedCaipChainId,
+  CommonNetworkConfiguration,
+  NonEvmNetworkConfiguration,
+  EvmNetworkConfiguration,
+  MultichainNetworkConfiguration,
+  MultichainNetworkControllerState,
+  MultichainNetworkControllerGetStateAction,
+  MultichainNetworkControllerSetActiveNetworkAction,
+  MultichainNetworkControllerStateChange,
+  MultichainNetworkControllerNetworkDidChangeEvent,
+  MultichainNetworkControllerActions,
+  MultichainNetworkControllerEvents,
+  MultichainNetworkControllerMessenger,
+} from './types';
+export {
+  checkIfSupportedCaipChainId,
+  toMultichainNetworkConfiguration,
+  toMultichainNetworkConfigurationsByChainId,
+  toEvmCaipChainId,
+} from './utils';